# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Get Main",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "main"
                ],
                "summary": "Get Main",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CodeResponse"
                        }
                    }
                }
            }
        },
        "/short/{key}": {
            "get": {
                "description": "Redirecting from short link to full",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shorter"
                ],
                "summary": "Short Link Redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short link key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
        "/shorter": {
            "post": {
                "description": "Getting full link and make it short",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shorter"
                ],
                "summary": "Do Link Short",
                "parameters": [
                    {
                        "type": "string",
                        "description": "full link",
                        "name": "full_link",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                }
            }
        },
        "controllers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}